Detect curves with theta turning recognition
As a practice test, theta turning recognition is used for curve detection. Theta turning recognition was proposed by Tsukasaki et al. in 2019 as a performant way to detect turnings, however it can only detect curves which have permanent turning motion. If the motion gets too weak for just a second during a curve (because of wind, GPS inaccuracy, … ), the curve can't be discovered.